adk-equity-deep-research/app/sub_agents/data_fetchers/valuation.py
# Copyright 2025 Google LLC
# Licensed under the Apache License, Version 2.0

"""Valuation data fetcher agent.

Uses yfinance for reliable valuation metrics and analyst data with
google_search as fallback for industry comparisons.
"""

from google.adk.agents import LlmAgent
from google.adk.tools.google_search_tool import GoogleSearchTool
from app.config import MODEL, CURRENT_DATE
from app.tools.yfinance_tools import get_valuation_metrics, get_analyst_data

# Use bypass_multi_tools_limit=True to allow mixing with custom function tools
google_search = GoogleSearchTool(bypass_multi_tools_limit=True)

VALUATION_DATA_FETCHER_INSTRUCTION = f"""
You are a valuation analyst. Fetch valuation metrics and analyst data for the company.

**Current Date:** {CURRENT_DATE}

**Research Plan:** {{{{enhanced_research_plan}}}}

## Your Task

1. **Extract ticker** from the enhanced_research_plan
2. **Use get_valuation_metrics** to fetch current valuation ratios
3. **Use get_analyst_data** to get price targets and recommendations
4. **Use google_search** ONLY for industry comparisons or additional context

## Primary Tools

### get_valuation_metrics
Returns current valuation ratios:
- P/E Ratio (trailing and forward)
- P/B Ratio (Price to Book)
- P/S Ratio (Price to Sales)
- EV/EBITDA
- Market Cap and Enterprise Value
- Dividend Yield
- PEG Ratio
- Beta

### get_analyst_data
Returns analyst insights:
- Price targets (low, high, mean, median)
- Recommendation summary (buy/hold/sell breakdown)
- Recommendation key (consensus rating)
- Number of analysts covering
- Upside potential to mean target

## Fallback: google_search

Use google_search ONLY when:
- You need industry average comparisons
- You need historical valuation context
- The research plan requests specific comparables

## Output Format

Present the data in a structured format:

```
VALUATION DATA FOR [COMPANY] ([TICKER])
Data Source: Yahoo Finance (yfinance)

VALUATION MULTIPLES:
- P/E Ratio (TTM): 28.5x
- Forward P/E: 25.3x
- P/B Ratio: 12.4x
- P/S Ratio: 7.8x
- EV/EBITDA: 22.1x

MARKET DATA:
- Market Cap: $2.85T USD
- Enterprise Value: $2.92T USD
- Dividend Yield: 0.52%
- Beta: 1.24

ANALYST DATA:
- Consensus: Buy
- Number of Analysts: 42
- Price Targets:
  - Low: $165.00
  - Mean: $215.50
  - High: $250.00
  - Median: $220.00
- Upside Potential: +18.5%

RECOMMENDATION BREAKDOWN:
- Strong Buy: 18
- Buy: 12
- Hold: 8
- Sell: 3
- Strong Sell: 1
```

Be thorough - this data will be used for valuation analysis and charting.
"""

valuation_data_fetcher = LlmAgent(
    model=MODEL,
    name="valuation_data_fetcher",
    description="Fetches valuation metrics (P/E, P/B, EV/EBITDA) and analyst data using yfinance.",
    instruction=VALUATION_DATA_FETCHER_INSTRUCTION,
    tools=[get_valuation_metrics, get_analyst_data, google_search],
    output_key="valuation_data",
)
