adk-equity-deep-research/requirements.txt
# =============================================================================
# Financial Data Visualization Agent - Requirements
# =============================================================================
# IMPORTANT: Requires Python 3.10+ (ADK v1.19+ dropped Python 3.9 support)
#
# If you get "Could not find a version that satisfies the requirement google-adk"
# try these steps:
#   1. Ensure Python 3.10+: python --version
#   2. Clear pip cache: pip cache purge
#   3. Upgrade pip: pip install --upgrade pip
#   4. Install from GitHub if PyPI fails:
#      pip install git+https://github.com/google/adk-python.git@main
# =============================================================================

# Core ADK dependencies (latest: v1.21.0 as of Dec 2025)
google-adk>=1.17.0
google-cloud-aiplatform>=1.112.0

# Data visualization (for local development/testing)
matplotlib
pandas
numpy

# Financial data from Yahoo Finance (structured, reliable data)
yfinance>=0.2.40

# Structured outputs
pydantic

# Vertex AI client
vertexai

# PDF export (optional - for generating PDF reports)
weasyprint>=60.0
